+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Enh #370: Refactor `Schema::normalizeDefaultValue()` method and move it to `ColumnFactory` class (@Tigrov)
 - New #373: Override `QueryBuilder::prepareBinary()` method (@Tigrov)
 - Chg #375: Update `QueryBuilder` constructor (@Tigrov)
+- Enh #374: Use `ColumnDefinitionBuilder` to generate table column SQL representation (@Tigrov)
 
 ## 1.3.0 March 21, 2024